More fixes for HTTP headers.
to address: HTTP requires that the ETags for two different responses associated with the same URI be different as well, to help caches and other receivers disambiguate them.
(so means we needs to call customGZipHandlerStart before customCacheControl)

and: HTTP requires 304 Not Modified responses to have certain headers, if they are also present in a normal (e.g., 200 OK response). ... This response is missing the following headers: last-modified.

via http://redbot.org/