Allow for the larger uploads
The juppy upload handler was never updated to handle larger than 640px uploads, so the upload manager would just use the default of 640px. 
Juppy doesnt yet have a option to select the size (although it can be set to upload larger than 640px images), so the best we can do is use the default set on user profile. 
... also had it use 1024 as default. The user can still explicitly set their default to 640px to override, or can just have juppy resize to 640px before upload.