Restructure Wikimedia Commons transfer section of "reuse" page
It now explains that the link to Commons will fill in everything but
categories.  The download link and the template box are still present
because they're sometimes useful for odd corner cases, and they might be
useful if the magic link stops working.

Also remove the link to geograph2commons, since it hasn't been working
for months and it doesn't offer any significant advantage over the new
magic link anyway.