Fix encoding issue with sending emails. THe plain text portion, is actully 'quoted-printable' and needs encoding as such.
Also fixed a secondary issue the rudiementory quoted-printable decoder, mishandled line breaks - the wrapping could of been in the middle of HTML tags.

Finally a third issue the smarty plugins/modifiers are encoded in the cache file. because the modifier was present some of the time when rendering ecard.tpl, it would fail. Needs to always be available (because could of been available at rendering time)
... fixes pretty rare but unlikey to be seen on production, as unlikelt that ecard.tpl would be 'first' compiled, during the send process.