Mostly functional database dump
Intended to be used to dumping an entire database, but does specifcally ignore Geograph Material Views

Note, its intended for dumping a temporally & non-active instance (eg restored from a snapshot temporarily), no attempt is made to make sure inconsistent snapshot. Locking and/or transactions are NOT used.
(we used it with a wrapper script that created an instance from shapshot, exported all tables, then deleted the instance)