First draft of a new responsive Reuse page.
Incorperates functions from More Sizes (select size to download) as well as Stamping options.

This is the first step in removing more.php and the form from stamp.php, both to simplify the site, but also to have less pages to crawl.
The three pages as set had a lot of overlapping and duplicate functionality.

They haven't been redirected yet, but will be.