Convert resizeImage to a WebWorker
The main reason is so can display progress of conversion, as processing happens in seperate thread, so main window can show messages
As a webworker needed to rearcitect to use OffscreenCanvas and ImageBitmap.
A happy side effect, OffscreenCanvas doesnt have a getDataURL so have to read as a Blob so can check size before adding exif! (if already too big, no point adding this time round!)